+++ Tests/Functional/Services/Persister/JobPersisterTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace webignition\BasilWorker\PersistenceBundle\Tests\Functional\Services\Persister;
 
 use webignition\BasilWorker\PersistenceBundle\Entity\Job;
-use webignition\BasilWorker\PersistenceBundle\Services\JobStore;
+use webignition\BasilWorker\PersistenceBundle\Services\Store\JobStore;
 use webignition\BasilWorker\PersistenceBundle\Services\Persister\JobPersister;
 use webignition\BasilWorker\PersistenceBundle\Tests\Functional\AbstractFunctionalTest;
