+++ Tests/Functional/Services/Store/SourceStoreTest.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace webignition\BasilWorker\PersistenceBundle\Tests\Functional\Services;
+namespace webignition\BasilWorker\PersistenceBundle\Tests\Functional\Services\Store;
 
 use webignition\BasilWorker\PersistenceBundle\Entity\Source;
 use webignition\BasilWorker\PersistenceBundle\Services\Factory\SourceFactory;
-use webignition\BasilWorker\PersistenceBundle\Services\SourceStore;
+use webignition\BasilWorker\PersistenceBundle\Services\Store\SourceStore;
 use webignition\BasilWorker\PersistenceBundle\Tests\Functional\AbstractFunctionalTest;
 
 class SourceStoreTest extends AbstractFunctionalTest
@@ -24,9 +24,9 @@ class SourceStoreTest extends AbstractFunctionalTest
             $this->sourceStore = $sourceStore;
         }
 
-        $sourceFactory = $this->container->get(\webignition\BasilWorker\PersistenceBundle\Services\Factory\SourceFactory::class);
-        self::assertInstanceOf(\webignition\BasilWorker\PersistenceBundle\Services\Factory\SourceFactory::class, $sourceFactory);
-        if ($sourceFactory instanceof \webignition\BasilWorker\PersistenceBundle\Services\Factory\SourceFactory) {
+        $sourceFactory = $this->container->get(SourceFactory::class);
+        self::assertInstanceOf(SourceFactory::class, $sourceFactory);
+        if ($sourceFactory instanceof SourceFactory) {
             $this->sourceFactory = $sourceFactory;
         }
     }
